+++ src/System.Buffers.Experimental/System/Buffers/BufferExtensions.cs
@@ -126,7 +126,7 @@ namespace System.Buffers
 
         // span creation helpers:
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static long IndexOf(this IMemorySegment<byte> sequence, ReadOnlySpan<byte> value)
+        public static long IndexOf(this IMemoryList<byte> sequence, ReadOnlySpan<byte> value)
         {
             var first = sequence.Memory.Span;
             var index = first.IndexOf(value);
@@ -138,7 +138,7 @@ namespace System.Buffers
             return IndexOfStraddling(first, sequence.Rest, value);
         }
 
-        public static long IndexOf(this IMemorySegment<byte> sequence, byte value)
+        public static long IndexOf(this IMemoryList<byte> sequence, byte value)
         {
             var first = sequence.Memory.Span;
             long index = first.IndexOf(value);
@@ -159,7 +159,7 @@ namespace System.Buffers
         // TODO (pri 3): I am pretty sure this whole routine can be written much better
 
         // searches values that potentially straddle between first and rest
-        internal static long IndexOfStraddling(this ReadOnlySpan<byte> first, IMemorySegment<byte> rest, ReadOnlySpan<byte> value)
+        internal static long IndexOfStraddling(this ReadOnlySpan<byte> first, IMemoryList<byte> rest, ReadOnlySpan<byte> value)
         {
             Debug.Assert(first.IndexOf(value) == -1);
             if (rest == null) return -1;
