+++ src/System.Buffers.Experimental/System/Buffers/Native/BufferManager.cs
