+++ src/System.Buffers.Experimental/System/Buffers/ReadOnlyBytes.cs
@@ -13,7 +13,7 @@ namespace System.Buffers
     public readonly struct ReadOnlyBytes : ISequence<ReadOnlyMemory<byte>>
     {
         readonly ReadOnlyMemory<byte> _first; // pointer + index:int + length:int
-        readonly IMemorySegment<byte> _all;
+        readonly IMemoryList<byte> _all;
 
         // For multi-segment ROB, this is the total length of the ROB
         // For single-segment ROB that are slices, this is the offset of the first byte from the original ROB
@@ -33,7 +33,7 @@ namespace System.Buffers
             _totalLengthOrOffset = offset;
         }
 
-        public ReadOnlyBytes(IMemorySegment<byte> segments, long length)
+        public ReadOnlyBytes(IMemoryList<byte> segments, long length)
         {
             // TODO: should we skip all empty buffers, i.e. of _first.IsEmpty?
             _first = segments.Memory;
@@ -41,7 +41,7 @@ namespace System.Buffers
             _totalLengthOrOffset = length;
         }
 
-        private ReadOnlyBytes(ReadOnlyMemory<byte> first, IMemorySegment<byte> all, long length)
+        private ReadOnlyBytes(ReadOnlyMemory<byte> first, IMemoryList<byte> all, long length)
         {
             // TODO: add assert that first overlaps all (once we have Overlap on Span)
             _first = first;
@@ -63,7 +63,7 @@ namespace System.Buffers
                 return false;
             }
 
-            var (segment, index) = position.Get<IMemorySegment<byte>>();
+            var (segment, index) = position.Get<IMemoryList<byte>>();
 
             if (segment == null)
             {
@@ -112,7 +112,7 @@ namespace System.Buffers
 
         public ReadOnlyMemory<byte> First => _first;
 
-        internal IMemorySegment<byte> Rest => _all?.Rest;
+        internal IMemoryList<byte> Rest => _all?.Rest;
 
         public long Length => _all == null ? _first.Length : _totalLengthOrOffset;
 
@@ -261,7 +261,7 @@ namespace System.Buffers
             return CursorAt(Rest, index - firstLength);
         }
 
-        private static Cursor CursorOf(IMemorySegment<byte> list, byte value)
+        private static Cursor CursorOf(IMemoryList<byte> list, byte value)
         {
             ReadOnlySpan<byte> first = list.Memory.Span;
             int index = first.IndexOf(value);
@@ -270,7 +270,7 @@ namespace System.Buffers
             return CursorOf(list.Rest, value);
         }
 
-        private static Cursor CursorAt(IMemorySegment<byte> list, int index)
+        private static Cursor CursorAt(IMemoryList<byte> list, int index)
         {
             if (list == null) return default;
             ReadOnlySpan<byte> first = list.Memory.Span;
@@ -320,100 +320,12 @@ namespace System.Buffers
         public SequenceEnumerator<ReadOnlyMemory<byte>, ReadOnlyBytes> GetEnumerator()
             => new SequenceEnumerator<ReadOnlyMemory<byte>, ReadOnlyBytes>(this);
 
-        class BufferListNode : IMemorySegment<byte>
-        {
-            internal Memory<byte> _data;
-            internal BufferListNode _next;
-            private long _runningIndex;
-
-            public BufferListNode(Memory<byte> data)
-            {
-                _data = data;
-                _next = null;
-                _runningIndex = 0;
-            }
-
-            private BufferListNode(Memory<byte> data, long runningIndex)
-            {
-                _data = data;
-                _runningIndex = runningIndex;
-            }
-
-            public BufferListNode Append(Memory<byte> data)
-            {
-                if (_next != null) throw new InvalidOperationException("Node cannot be appended");
-                var node = new BufferListNode(data, _runningIndex + Length);
-                _next = node;
-                return node;
-            }
-
-            public Memory<byte> Memory => _data;
-            public IMemorySegment<byte> Rest => _next;
-
-            public long Length => _data.Length;
-
-            public long VirtualIndex => _runningIndex;
-
-            public int CopyTo(Span<byte> buffer)
-            {
-                int copied = 0;
-                Position position = default;
-                var free = buffer;
-                while (TryGet(ref position, out ReadOnlyMemory<byte> segment, true))
-                {
-                    if (segment.Length > free.Length)
-                    {
-                        segment.Span.Slice(0, free.Length).CopyTo(free);
-                        copied += free.Length;
-                    }
-                    else
-                    {
-                        segment.Span.CopyTo(free);
-                        copied += segment.Length;
-                    }
-                    free = buffer.Slice(copied);
-                    if (free.Length == 0) break;
-                }
-                return copied;
-            }
-
-            public bool TryGet(ref Position position, out ReadOnlyMemory<byte> item, bool advance = true)
-            {
-                var result = TryGet(ref position, out Memory<byte> memory, advance);
-                item = memory;
-                return result;
-            }
-
-            public bool TryGet(ref Position position, out Memory<byte> item, bool advance = true)
-            {
-                if (position == default)
-                {
-                    item = _data;
-                    if (advance)
-                    {
-                        position.SetItem(_next);
-                    }
-                    return (!_data.IsEmpty || _next != null);
-                }
-                else if (position.IsEnd)
-                {
-                    item = default;
-                    return false;
-                }
-
-                var sequence = position.GetItem<BufferListNode>();
-                item = sequence._data;
-                if (advance) { position.SetItem(sequence._next); }
-                return true;
-            }
-        }
-
         public static ReadOnlyBytes Create(params byte[][] buffers)
         {
             if (buffers.Length == 0) return Empty;
             if (buffers.Length == 1) return new ReadOnlyBytes(buffers[0]);
 
-            BufferListNode first = new BufferListNode(buffers[0]);
+            MemoryList first = new MemoryList(buffers[0]);
             var last = first;
             for(int i=1; i<buffers.Length; i++)
             {
@@ -425,10 +337,10 @@ namespace System.Buffers
 
         public readonly struct Cursor
         {
-            internal readonly IMemorySegment<byte> _node;
+            internal readonly IMemoryList<byte> _node;
             internal readonly int _index;
 
-            public Cursor(IMemorySegment<byte> node, int index)
+            public Cursor(IMemoryList<byte> node, int index)
             {
                 _node = node;
                 _index = index;
